requirements.txt
google-generativeai
gradio
pypdf2
faiss-cpu
numpy
python-dotenv